+++ CHANGELOG.md
@@ -10,6 +10,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Generated mock objects will no longer trigger Clippy's (off by default)
   `clippy::struct_field_names` lint.
   ([#691](https://github.com/asomers/mockall/pull/691))
+- Fixed `#[automock]` on methods whose closures use associated types.
+  ([#692](https://github.com/asomers/mockall/pull/692))
 
 ## [ 0.15.0 ] - 2026-06-28